Use custom babel version and check for reasonable CID

+++ cb/trunk/get_cid_cml.pl
@@ -24,11 +24,11 @@ my $sql = $db->prepare("SELECT cid FROM compounds WHERE cid != ''");
 $sql->execute();
 while (my $row = $sql->fetchrow_hashref()) {
   my $cid = $row->{"cid"};
-  if (!(-e "$imageDir".$cid.".cml")) {
+  if (!(-e "$imageDir".$cid.".cml") && $cid > 0) {
     print "Need to get a CML for CID $cid\n";
     my $url = "http://pubchem.ncbi.nlm.nih.gov/summary/summary.cgi?cid=$cid\&disopt=DisplaySDF";
     `wget -q -O $cid.mol "$url"`;
-    `babel $cid.mol -ocml $cid.cml`;
+    `/home/egonw/progs/bin/babel $cid.mol -ocml $cid.cml`;
     `cat $cid.cml | grep -v "<?xml" > $cid.out.cml`;
     `mv $cid.out.cml $imageDir$cid.cml`;
     `rm $cid.mol`;
